Rename files and update layout to fix hydration warning
-Renamed files to align with the updated naming conventions.
-Made changes to the layout structure to resolve hydration warnings.
