+++ src/app/layout.tsx
@@ -7,13 +7,13 @@ import ThemeRegistry from "@/context/ThemeRegistry/ThemeRegistry";
 const geistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
-  display: "swap", // Add `display: swap` for better font loading performance
+  display: "swap",
 });
 
 const geistMono = Geist_Mono({
   variable: "--font-geist-mono",
   subsets: ["latin"],
-  display: "swap", // Add `display: swap`
+  display: "swap",
 });
 
 export const metadata: Metadata = {
@@ -27,7 +27,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
+    <html lang="en" suppressHydrationWarning>
       <body className={`${geistSans.variable} ${geistMono.variable}`}>
         <ThemeRegistry>
           <ThemeContextProvider>{children}</ThemeContextProvider>
