+++ src/app/login/login-page.module.css
@@ -24,7 +24,6 @@
   justify-content: center;
   align-items: center;
   width: 50%;
-  /* height: 100%; */
   border-radius: 0.5rem;
   object-fit: contain;
 }
