+++ src/components/login/LoginImage.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import Image from "next/image";
